+++ YarnSpinner.Compiler/Compiler.cs
@@ -918,8 +918,7 @@ namespace Yarn.Compiler
                 {
                     Name = function.Key,
                     Type = functionType,
-                    SourceFileLine = -1,
-                    SourceNodeLine = -1,
+                    Range = { },
                     SourceFileName = Declaration.ExternalDeclaration,
                     SourceNodeName = null,
                 };
@@ -948,8 +947,8 @@ namespace Yarn.Compiler
             YarnSpinnerParser parser = new YarnSpinnerParser(tokens);
 
             // turning off the normal error listener and using ours
-            var parserErrorListener = new ParserErrorListener();
-            var lexerErrorListener = new LexerErrorListener();
+            var parserErrorListener = new ParserErrorListener(fileName);
+            var lexerErrorListener = new LexerErrorListener(fileName);
 
             parser.ErrorHandler = new ErrorStrategy();