+++ YarnSpinner.Compiler/TypeCheckVisitor.cs
@@ -207,9 +207,20 @@ namespace Yarn.Compiler
                     IsImplicit = true,
                     Description = $"Implicit declaration of function at {sourceFileName}:{functionCall.Start.Line}:{functionCall.Start.Column}",
                     SourceFileName = sourceFileName,
-                    SourceFileLine = functionCall.Start.Line,
                     SourceNodeName = currentNodeName,
-                    SourceNodeLine = functionCall.Start.Line - (this.currentNodeContext.BODY_START().Symbol.Line + 1),
+                    Range = new Range
+                    {
+                        Start =
+                        {
+                            Line = functionCall.Start.Line - 1,
+                            Character = functionCall.Start.Column,
+                        },
+                        End =
+                        {
+                            Line = functionCall.Stop.Line - 1,
+                            Character = functionCall.Stop.Column + functionCall.Stop.Text.Length,
+                        },
+                    },
                 };
 
                 // Create the array of parameters for this function based
@@ -269,7 +280,7 @@ namespace Yarn.Compiler
 
                 if (TypeUtil.IsSubType(expectedType, suppliedType) == false)
                 {
-                    this.diagnostics.Add(new Diagnostic(this.sourceFileName, functionCall, $"{functionName} parameter {i + 1} expects a {expectedType.Name}, not a {(suppliedType?.Name ?? "undefined")}"));
+                    this.diagnostics.Add(new Diagnostic(this.sourceFileName, functionCall, $"{functionName} parameter {i + 1} expects a {expectedType?.Name ?? "undefined"}, not a {suppliedType?.Name ?? "undefined"}"));
                     return functionType.ReturnType;
                 }
             }
@@ -305,12 +316,19 @@ namespace Yarn.Compiler
 
         public override Yarn.IType VisitSet_statement([NotNull] YarnSpinnerParser.Set_statementContext context)
         {
-            var expressionType = Visit(context.expression());
-            var variableType = Visit(context.variable());
+            var variableContext = context.variable();
+            var expressionContext = context.expression();
 
-            var variableName = context.variable().GetText();
+            if (expressionContext == null || variableContext == null) {
+                return BuiltinTypes.Undefined;
+            }
 
-            ParserRuleContext[] terms = { context.variable(), context.expression() };
+            var expressionType = base.Visit(expressionContext);
+            var variableType = base.Visit(variableContext);
+
+            var variableName = variableContext.GetText();
+
+            ParserRuleContext[] terms = { variableContext, expressionContext };
 
             Yarn.IType type;
 
@@ -341,13 +359,24 @@ namespace Yarn.Compiler
                         var decl = new Declaration
                         {
                             Name = variableName,
-                            Description = $"{System.IO.Path.GetFileName(sourceFileName)}, node {currentNodeName}, line {context.Start.Line - nodePositionInFile}",
+                            Description = $"Implicitly declared in {System.IO.Path.GetFileName(sourceFileName)}, node {currentNodeName}",
                             Type = expressionType,
                             DefaultValue = DefaultValueForType(expressionType),
                             SourceFileName = sourceFileName,
-                            SourceFileLine = context.Start.Line,
                             SourceNodeName = currentNodeName,
-                            SourceNodeLine = context.Start.Line - nodePositionInFile,
+                            Range = new Range
+                            {
+                                Start =
+                                {
+                                    Line = variableContext.Start.Line - 1,
+                                    Character = variableContext.Start.Column,
+                                },
+                                End =
+                                {
+                                    Line = variableContext.Stop.Line - 1,
+                                    Character = variableContext.Stop.Column + variableContext.GetText().Length,
+                                },
+                            },
                             IsImplicit = true,
                         };
                         NewDeclarations.Add(decl);
@@ -480,17 +509,13 @@ namespace Yarn.Compiler
                 .Concat(terms.OfType<YarnSpinnerParser.ValueVarContext>().Select(v => v.variable()))
                 .Where(c => c != null);
 
-            // Get their names
-            var variableNames = variableContexts
-                .Select(v => v.VAR_ID().GetText())
+            // Build the list of variable contexts that we don't have a
+            // declaration for. We'll check for explicit declarations first.
+            var undefinedVariableContexts = variableContexts
+                .Where(v => Declarations.Any(d => d.Name == v.VAR_ID().GetText()) == false)
                 .Distinct();
 
-            // Build the list of variable names that we don't have a
-            // declaration for. We'll check for explicit declarations first.
-            var undefinedVariableNames = variableNames
-                .Where(name => Declarations.Any(d => d.Name == name) == false);
-
-            if (undefinedVariableNames.Count() > 0)
+            if (undefinedVariableContexts.Count() > 0)
             {
                 // We have references to variables that we don't have a an
                 // explicit declaration for! Time to create implicit
@@ -502,19 +527,30 @@ namespace Yarn.Compiler
                 // The start line of the body is the line after the delimiter
                 int nodePositionInFile = this.currentNodeContext.BODY_START().Symbol.Line + 1;
 
-                foreach (var undefinedVariableName in undefinedVariableNames)
+                foreach (var undefinedVariableContext in undefinedVariableContexts)
                 {
                     // Generate a declaration for this variable here.
                     var decl = new Declaration
                     {
-                        Name = undefinedVariableName,
-                        Description = $"{System.IO.Path.GetFileName(sourceFileName)}, node {currentNodeName}, line {positionInFile - nodePositionInFile}",
+                        Name = undefinedVariableContext.VAR_ID().GetText(),
+                        Description = $"Implicitly declared in {System.IO.Path.GetFileName(sourceFileName)}, node {currentNodeName}",
                         Type = expressionType,
                         DefaultValue = DefaultValueForType(expressionType),
                         SourceFileName = sourceFileName,
-                        SourceFileLine = positionInFile,
                         SourceNodeName = currentNodeName,
-                        SourceNodeLine = positionInFile - nodePositionInFile,
+                        Range = new Range
+                        {
+                            Start =
+                            {
+                                Line = undefinedVariableContext.Start.Line - 1,
+                                Character = undefinedVariableContext.Start.Column,
+                            },
+                            End =
+                            {
+                                Line = undefinedVariableContext.Stop.Line - 1,
+                                Character = undefinedVariableContext.Stop.Column + undefinedVariableContext.Stop.Text.Length,
+                            },
+                        },
                         IsImplicit = true,
                     };
                     NewDeclarations.Add(decl);